refactor(library): cache and pipeline are now packages
This diff transforms cache and pipeline to become packages
instead of modules. This change opens us the possibility
to reorganize and refactor the code more easily.
